refactor: Remove Pinecone Local Index and consolidate Pinecone index implementation
- Deleted `pinecone_local.py` and `10-pinecone-local.ipynb`
- Updated `pinecone.py` to handle both cloud and local Pinecone index scenarios
- Modified test files to use `PineconeIndex` instead of `PineconeLocalIndex`
- Updated GitHub Actions workflow to simplify Python version testing
- Improved error handling and host URL management in Pinecone index methods
